回复问题异常处理后端接口

+++ src/main/java/com/yao/forum/Controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getQuestionById(id);
         model.addAttribute("title",question.getTitle());
@@ -49,7 +49,7 @@ public class PublishController {
                             @RequestParam("title")  String title,
                             @RequestParam("description")  String description,
                             @RequestParam("tag")  String tag,
-                            @RequestParam("id") Integer id
+                            @RequestParam("id") Long id
                             ){
         model.addAttribute("title",title);
         model.addAttribute("description",description);
