+++ src/main/java/com/yao/forum/Controller/QuestionController.java
@@ -14,7 +14,7 @@ public class QuestionController {
     QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String quetions(@PathVariable(name = "id") Integer id,
+    public String quetions(@PathVariable(name = "id") Long id,
                            Model model){
 
         QuestionDTO questionDTO=questionService.getQuestionById(id);
